+++ modin/pandas/api/extensions/test/test_extension_pd.py
@@ -12,7 +12,7 @@
 # governing permissions and limitations under the License.
 
 import modin.pandas as pd
-from modin.extensions import register_pd_accessor
+from modin.pandas.api.extensions import register_pd_accessor
 
 
 def test_dataframe_extension_simple_method():
